+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Release (Unreleased)
+
+* `github.com/patrickcping/pingone-go-sdk-v2/management` : [v0.55.0](./management/CHANGELOG.md)
+  * **Enhancement** Support `AlternativeIdentifiers` on the `Population` data model.
+  * **Enhancement** Support `PreferredLanguage` on the `Population` data model.
+  * **Enhancement** Support `Theme` on the `Population` data model.
+
 # Release (2025-04-28)
 
 * `github.com/patrickcping/pingone-go-sdk-v2` : v0.12.14